requirements.txt
absl-py
aiohappyeyeballs
aiohttp
aiosignal
asttokens
attrs
audioread
certifi
cffi
charset-normalizer
colorama
comm
contourpy
cycler
datasets
debugpy
decorator
dill
editdistance
executing
filelock
flashlight-text
fonttools
frozenlist
fsspec
grpcio
huggingface-hub
idna
ipykernel
ipython
jedi
Jinja2
joblib
jupyter_client
jupyter_core
kiwisolver
lazy_loader
librosa
llvmlite
Markdown
MarkupSafe
matplotlib
matplotlib-inline
mpmath
msgpack
multidict
multiprocess
nest-asyncio
networkx
numba
numpy
packaging
pandas
parso
pillow
platformdirs
pooch
prompt_toolkit
protobuf
psutil
pure_eval
pyarrow
pycparser
Pygments
pyparsing
python-dateutil
pytz
pywer
#pywin32
PyYAML
pyzmq
requests
scikit-learn
scipy
seaborn
setuptools
six
soundfile
soxr
stack-data
sympy
tensorboard
tensorboard-data-server
threadpoolctl
#torch
#torchaudio
#torchvision
torchinfo
tornado
tqdm
traitlets
transformers
typing_extensions
tzdata
urllib3
wcwidth
Werkzeug
xxhash
yarl
